Refactor: split Completions.scala into smaller files
related: https://github.com/scalameta/metals/issues/672
